CorpMSP
Over 480 GB and 75,000 files, emails and disk images from CorpMSP, a federal institution providing support to small and medium-sized businesses. The controlling shareholder of CorpMSP is the Russian Federation.

According to NB65, CorpMSP is also allegedly used as a shell company for digital spies, and maintains contracts with the Ministry of Defense. The files which documented these allegations were quickly removed from Mega, which allegedly hosted NB65's only copy. According to NB65, the 9GB of files included client contracts and image files of Ministry of Defense contracts.

https://twitter.com/xxNB65/status/1518594916431060992/photo/1
https://twitter.com/xxNB65/status/1519772298282254341
https://twitter.com/xxNB65/status/1520037618670526467