Odebrecht
390,000 application (non-employee) emails ranging from 2008 to 2022 and 365 fserver emails ranging from April 2021 to early May 2022 from Novonor, previously Odebrecht S.A.

The fserver emails contain details of SWIFT transfers to their lawyers in the United States, as well as their letter of engagement. The emails also reveal SWIFT transfers of over $100,000 to employees in the United States from the Brazilian company's Dominican Republic-based bank account to the Miami-based employee's Brazilian-based bank account.
https://en.wikipedia.org/wiki/Novonor
https://en.wikipedia.org/wiki/Odebrecht_Case
https://en.wikipedia.org/wiki/Clifford_Chance