Hunter Biden emails
Approximately 128,500 emails allegedly from the Hunter Biden laptop. The emails are primarily dated between 2009 and 2019. There are some anomalies with the dates and the number of emails, in addition to concerns about the chain of custody. There are considerable issues with this dataset including signs of tampering, as well as misconceptions about the contents. See below.

More than one altered or implanted emails was detected in a version of this dataset, which is being distributed by Trump allies and former staffers. The known possibly tampered emails were created between August 31, 2020 and September 2, 2020. The existence of other possibly tampered emails cannot be ruled out. See also: Timeline disecrepancies below.(https://ddosecrets.com/wiki/Hunter_Biden_emails#Timeline_discrepancies)

We present the dataset now in full context to counteract possible deceptions by persons with an agenda who are currently distributing the dataset without the relevant context or warnings.